\name{boolfirst}
\alias{boolfirst}
\title{ Deprecated Function }
\description{
  This function was implemented in previous versions, but is
  now deprecated. For an improved implementation, see the
  \code{sim} function in the Zelig library.
}
\usage{
  boolfirst(...)
}
\arguments{
  \item{...}{It does not matter.}
}
\value{
  A message indicating deprecation is printed; nothing is returned.
}
\references{ Braumoeller, Bear F. (2003) "Causal Complexity and the
     Study of Politics." \emph{Political Analysis} 11(3): 209-233. }

\author{
Bear F. Braumoeller, Harvard University, \email{bfbraum@fas.harvard.edu}, \cr
Ben Goodrich, Harvard University, \email{goodrich@fas.harvard.edu}, and \cr
Jacob Kline, Harvard University, \email{jkline@fas.harvard.edu}
}

\seealso{ \code{\link{boolean}}, \code{\link{boolprep}}, 
     and \code{\link{boolprof}}. }

\keyword{logic}
\keyword{optimize}
\keyword{distribution}
\keyword{models}
