\name{coef.boolean}
\docType{methods}
\alias{coef.boolean}
\alias{coef-methods}
\alias{coef,ANY-method}
\alias{coef,booltest-method}
\title{Coefficients for Boolean Logit and Probit Models}
\description{
'coef' method for objects of class '"booltest"'
}
\section{Usage}{\describe{

\item{object = "booltest"}{coef(object)}
}}
\keyword{methods}
