\name{plot.boolean}
\docType{methods}
\alias{plot.boolean}
\alias{plot-methods}
\alias{plot,ANY-method}
\alias{plot,booltest-method}
\title{Plot Matrix for Boolean Logit and Probit Models}
\description{
'plot' method for objects of class '"booltest"'
}
\section{Usage}{\describe{
\item{object = "booltest"}{plot(object, panel = "boolfirst")}
}}
\arguments{
  \item{panel}{The individual graphs to be drawn in the plot matrix.
  Options are "\code{\link{boolfirst}}" for first differences and
  "\code{\link{boolprof}}" for profile likelihoods.}
}
\keyword{methods}

