\name{print.summary.booltest}
\docType{methods}
\alias{print.summary.booltest}
\alias{print-methods}
\alias{print,ANY-method}
\alias{print,summary.booltest-method}
\title{'print' Method for Boolean Binary Response Model Summaries}
\description{
'print' method for objects of class '"summary.booltest"'
}
\section{Usage}{\describe{
\item{object = "booltest"}{print(summary(object))}
}}
\keyword{methods}

