\name{show.boolean}
\docType{methods}
\alias{show.boolean}
\alias{show-methods}
\alias{show,ANY-method}
\alias{show,booltest-method}
\title{'show' Method for Boolean Logit and Probit Models}
\description{
'show' method for objects of class '"booltest"'
}
\section{Usage}{\describe{
\item{object = "booltest"}{show(object)}
}}
\keyword{methods}

