\name{summary.boolean}
\docType{methods}
\alias{summary.boolean}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,booltest-method}
\title{'summary' Method for Boolean Logit and Probit Models}
\description{
'summary' method for objects of class '"booltest"'
}
\section{Usage}{\describe{
\item{object = "booltest"}{summary(object)}
}}
\keyword{methods}

