\name{print.summary.booltest}
\alias{print.summary.booltest}
\alias{print}
\title{'print' Method for Boolean Binary Response Model Summaries}
\description{
'print' method for objects of class 'summary.booltest'
}

\usage{
## S3 method for class 'summary.booltest'
\method{print}{summary.booltest}(x, ...)
}

\arguments{
  \item{x}{An object of class 'summary.booltest'}
  \item{...}{Further arguments.}
}
\value{
  Does not return anything.
}
\references{
  Braumoeller, Bear F. (2003) "Causal Complexity and the Study
  of Politics." \emph{Political Analysis} 11(3): 209-233.
}
\author{
Bear F. Braumoeller, Harvard University, \email{braumoeller.1@polisci.osu.edu},\cr
Ben Goodrich, Harvard University, \email{goodrich@fas.harvard.edu}, and\cr
Jacob Kline, Harvard University, \email{jkline@fas.harvard.edu}
}

\keyword{methods}



